Remove usage of CustomButton widgets
When a modal window is shown, and a CustomButton widget has focus at the
same time then the widget still assumes, that it has the focus. To solve
this issue without dirty hacks, own ImageButton widgets are now used
instead of CustomButton widgets. The FileUploadImageButton widget is
introduced to allow use FileUpload as a ImageButton widget.